backend/requirements.txt
requests
beautifulsoup4
voyageai
supabase
python-dotenv
anthropic
pydantic>=2.0
fastapi
uvicorn[standard]
httpx
langgraph
